module/sotest:use Compile dynamic modules using a unified compilation method
