--- examples/module/drivers/.gitignore
@@ -1,5 +1,0 @@
-/fsroot
-/romfs.c
-/cromfs.c
-/romfs.img
-/mod_symtab.c
