--- examples/module/drivers/chardev/.gitignore
@@ -1,1 +1,0 @@
-/chardev
