+++ examples/module/main/Kconfig
