--- examples/sotest/lib/.gitignore
@@ -1,4 +1,0 @@
-/fsroot
-/romfs.c
-/romfs.img
-/sot_symtab.c
