--- examples/sotest/lib/modprint/.gitignore
@@ -1,1 +1,0 @@
-/modprint
