--- examples/sotest/lib/sotest/.gitignore
@@ -1,1 +1,0 @@
-/sotest
