+++ examples/sotest/main/.gitignore
@@ -1,0 +1,3 @@
+/sotest_symtab.c
+/sotest_romfs.c
+/sotest_romfs.img
