+++ examples/sotest/main/Kconfig
