+++ examples/sotest/main/sotest_main.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * apps/examples/sotest/sotest_main.c
+ * apps/examples/sotest/main/sotest_main.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
