+++ examples/sotest/sotest/Make.defs
@@ -1,5 +1,5 @@
 ############################################################################
-# apps/examples/sotest/lib/modprint/Makefile
+# apps/examples/sotest/sotest/Make.defs
 #
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
@@ -18,47 +18,6 @@
 #
 ############################################################################
 
-include $(APPDIR)/Make.defs
-
-ifeq ($(CONFIG_EXAMPLES_SOTEST_LIBGCC),y)
-LIBGCC = "${shell $(CC) $(ARCHCPUFLAGS) -print-libgcc-file-name 2>/dev/null}"
-ifneq ($(LIBGCC),)
-  LDLIBPATH += -L "${shell dirname $(LIBGCC)}"
-  LDLIBS += -lgcc
+ifneq ($(CONFIG_EXAMPLES_SOTEST),)
+CONFIGURED_APPS += $(APPDIR)/examples/sotest/sotest
 endif
-endif
-
-ifeq ($(CONFIG_EXAMPLES_SOTEST_LIBC),y)
-LDMODULEFLAGS += -Bstatic
-LDLIBPATH +=  -L $(NUTTXLIB)
-endif
-
-ifeq ($(CONFIG_EXAMPLES_SOTEST_LIBC),y)
-LDLIBS += -lc
-endif
-
-BIN = modprint
-
-SRCS = $(BIN).c
-OBJS = $(SRCS:.c=$(OBJEXT))
-
-all: $(BIN)
-.PHONY: all clean install
-
-$(OBJS): %$(OBJEXT): %.c
-	@echo "MODULECC: $<"
-	$(Q) $(MODULECC) -c $(CMODULEFLAGS) $< -o $@
-
-$(BIN): $(OBJS)
-	@echo "MODULELD: $<"
-	$(Q) $(MODULELD) $(LDMODULEFLAGS) $(LDLIBPATH) -o $@ $(ARCHCRT0OBJ) $^ $(LDLIBS)
-
-$(FSROOT_DIR)/$(BIN): $(BIN)
-	$(Q) mkdir -p $(FSROOT_DIR)
-	$(Q) install $(BIN) $(FSROOT_DIR)/$(BIN)
-
-install: $(FSROOT_DIR)/$(BIN)
-
-clean:
-	$(call DELFILE, $(BIN))
-	$(call CLEAN)
